streamlit_app (copy)/requirements.txt
pytesseract
Pillow
PyYAML
PyMuPDF
opencv-python-headless
numpy
transformers
timm
pandas
openpyxl
openai
streamlit
scikit-learn
xlsxwriter
surya-ocr
lxml
pytesseract
Pillow
PyYAML
PyMuPDF
opencv-python-headless
numpy
transformers
timm
pandas
openpyxl
openai
streamlit
scikit-learn
xlsxwriter
docling
opencv-python
Ghostscript
ollama
openai
anthropic
tiktoken
faiss-cpu # Vector Store
sentence_transformers # Huggingface Embedding Models
langchain_ollama
langchain_huggingface
langchain_community
python-dotenv
PyPDF2
tesserocr
